1. 预下载、下载、缓冲大小、预加载配置的类型修改为浮点类型 2. 点播预加载、缓冲大小增加默认值 3. 规避在flutter高版本中，ios app在后台停留时间过长导致app被杀的时候，如果app还一直在离线下载，会造成app资源释放不一致问题，导致flutter播放器插件engine已经结束，但是下载进程还在回调，产生了crash (merge request !257)
Squash merge branch 'kongdywang/pro/fix_ios_download_crash_when_kill' into 'Professional'
1. 预下载、下载、缓冲大小、预加载配置的类型修改为浮点类型
2. 点播预加载、缓冲大小增加默认值
3. 规避在flutter高版本中，ios app在后台停留时间过长导致app被杀的时候，如果app还一直在离线下载，会造成app资源释放不一致问题，导致flutter播放器插件engine已经结束，但是下载进程还在回调，产生了crash

+++ Flutter/lib/Core/pigeons/txplayer_message.dart
@@ -110,10 +110,10 @@ class FTXVodPlayConfigPlayerMsg {
   int? progressInterval;
 
   // 最大播放缓冲大小，单位 MB。此设置会影响playableDuration，设置越大，提前缓存的越多
-  int? maxBufferSize;
+  double? maxBufferSize;
 
   // 预加载最大缓冲大小，单位：MB
-  int? maxPreloadSize;
+  double? maxPreloadSize;
 
   // 首缓需要加载的数据时长，单位ms，默认值为100ms
   int? firstStartPlayBufferTime;
@@ -263,7 +263,7 @@ class DoubleMsg {
 
 class PreLoadMsg {
   String? playUrl;
-  int? preloadSizeMB;
+  double? preloadSizeMB;
   int? preferredResolution;
 }
 
@@ -272,7 +272,7 @@ class PreLoadInfoMsg {
   String? fileId;
   String? pSign;
   String? playUrl;
-  int? preloadSizeMB;
+  double? preloadSizeMB;
   int? preferredResolution;
   int? tmpPreloadTaskId;
 }
